# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample library server",
        "title": "Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/addBook": {
            "post": {
                "description": "JWT authentication via cookie only for admin.\nThe JWT token should be stored in a cookie named \"jwt\".",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Add a new book",
                "parameters": [
                    {
                        "description": "Book Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AddBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/deleteBook": {
            "delete": {
                "description": "deletes the book from the library\nJWT authentication via cookie only for admin.\nThe JWT token should be stored in a cookie named \"jwt\".",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Delete the book",
                "parameters": [
                    {
                        "description": "Book Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getBook": {
            "get": {
                "description": "Get detailed information about a single book by ID\nJWT authentication via cookie.\nThe JWT token should be stored in a cookie named \"jwt\".",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Get one book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "bookId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getBooks": {
            "get": {
                "description": "Retrieve all books, optionally sorted by a specific field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Get list of books",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field to sort by (e.g., 'id')",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/modifyingBook": {
            "post": {
                "description": "JWT authentication via cookie only for admin.\nThe JWT token should be stored in a cookie named \"jwt\".\nJWT Bearer authentcation only admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "book"
                ],
                "summary": "Modifying book",
                "parameters": [
                    {
                        "description": "Book Data",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ModifyingBookRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "handlers.AddBookRequest": {
            "type": "object",
            "required": [
                "author",
                "genre",
                "published_year",
                "title"
            ],
            "properties": {
                "author": {
                    "description": "Автор",
                    "type": "string",
                    "example": "John Doe"
                },
                "description": {
                    "description": "Описание книги",
                    "type": "string",
                    "example": "Эта книга — идеальный выбор для тех, кто хочет начать свое путешествие в программировании на языке Go."
                },
                "genre": {
                    "description": "Жанра",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Учебная литература"
                    ]
                },
                "published_year": {
                    "description": "Год публикации",
                    "type": "string",
                    "example": "2024"
                },
                "title": {
                    "description": "Название книги",
                    "type": "string",
                    "example": "Golang Basics"
                }
            }
        },
        "handlers.DeleteBookRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "handlers.ModifyingBookRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "example": "Jeff Bezos"
                },
                "description": {
                    "type": "string",
                    "example": "Explore the ultimate question: Why is there something rather than nothing? This thought-provoking journey through philosophy, science, and metaphysics challenges readers to ponder existence itself, blending deep inquiry with accessible insight. A must-read for curious minds."
                },
                "genre": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Детектив"
                    ]
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "published_year": {
                    "type": "string",
                    "example": "2021"
                },
                "title": {
                    "type": "string",
                    "example": "Why Does the World Exist?"
                }
            }
        },
...
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Genre"
                    }
                },
                "published_year": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Genre": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Book"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}